requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.7  # required for image upload endpoints (UploadFile)
httpx>=0.27.0            # for async API testing with httpx.AsyncClient
pytest>=8.1.0
pytest-asyncio>=0.23.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0          # async PostgreSQL driver (Neon)
alembic>=1.13.0
python-dotenv>=1.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
firebase-admin>=6.5.0

# UI (Streamlit demo interface)
# streamlit>=1.32.0
# plotly>=5.18.0          # for confidence gauge chart in UI

# ML pipeline (install separately when training — TF is large)
# tensorflow>=2.16.0
# numpy>=1.26.0
# scikit-learn>=1.4.0    # for evaluation metrics
# matplotlib>=3.8.0      # for confusion matrix plot + Grad-CAM overlay
